Configure base.html.twig file
-Updated the base file to contain all necessary biolerplate code that all the pages require.

-Updated all the pages to work with the base.html.twig file.
