+++ public/js/home.js
@@ -1,0 +1,42 @@
+//Javascript Code
+
+//Navigation Bar
+window.addEventListener("scroll", function() {
+    var navbar = document.querySelector("nav");
+    navbar.classList.toggle("sticky", window.scrollY > 0);
+});
+
+const navSlide = () => {
+    const burger = document.querySelector('.burger');
+    const nav = document.querySelector('.nav-links');
+    const navLinks = document.querySelectorAll('.nav-links li');
+
+    //Toggle the Navigation
+    burger.addEventListener('click', () => {
+        nav.classList.toggle('nav-active');
+
+        //Animate the Links
+        navLinks.forEach((link, index) => {
+            if(link.style.animation) {
+                link.style.animation = '';
+            } else {
+                link.style.animation = `navLinkFade 0.5s ease forwards ${index / 7 + 0.4}s`
+            }
+        });
+
+        //Burger Animation
+        burger.classList.toggle('toggle');
+    });
+}
+
+navSlide();
+
+//Parallax Header Image Effect
+const parallax = document.getElementById("parallax");
+
+window.addEventListener("scroll", function() {
+    let offset = window.pageYOffset;
+    parallax.style.backgroundPositionY = offset * 0.1 + "px";
+});
+
+
