+++ templates/base.html.twig
@@ -1,12 +1,62 @@
 <!DOCTYPE html>
-<html>
+<html lang="en">
     <head>
+        <!-- META DATA -->
         <meta charset="UTF-8">
-        <title>{% block title %}Welcome!{% endblock %}</title>
-        {% block stylesheets %}{% endblock %}
+        <meta name="description" content="Creature is a South African Blogging Website for those interested in Wildlife!">
+        <meta name="keywords" content="Creature Blog, Wildlife, Nature, Blogging, South African Wildlife, Writing, Ideas, Thinking, Website">
+        <meta name="author" content="Dylan da Silva, Dylan Vanhout, Pfano Khampha">
+        <meta name="viewport" content="width=device-width, initial-scale=1.0">       
+        
+        <!-- WEBSITE TITLE -->
+        <title>{% block title %}Creature{% endblock %}</title>
+        <link rel="icon" href="{{ asset('img/colour-logo.ico') }}">
+        
+        <!-- JQUERY -->
+        <script src="https://code.jquery.com/jquery-3.5.1.min.js" integrity="sha256-9/aliU8dGd2tb6OSsuzixeV4y/faTqgFtohetphbbj0=" crossorigin="anonymous"></script>
+        
+        <!-- FONTS -->
+        <link href="https://fonts.googleapis.com/css2?family=Lato:wght@300;400;700&display=swap" rel="stylesheet">
+        <link href="https://fonts.googleapis.com/css2?family=Caveat+Brush&display=swap" rel="stylesheet">
+
+        <!-- BOOTSTRAP AND PAPER KIT-->
+        <link rel="stylesheet" href="{{ asset('css/paper-kit.css') }}">
+        <link rel="stylesheet" href="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/css/bootstrap.min.css" integrity="sha384-9gVQ4dYFwwWSjIDZnLEWnxCjeSWFphJiwGPXr1jddIhOegiu1FwO5qRGvFXOdJZ4" crossorigin="anonymous">   
+            
+        <!-- CSS -->
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/animate.css/4.1.1/animate.min.css">
+        <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/5.14.0/css/all.min.css">
+
+        {% block stylesheets %}
+            <link rel="stylesheet" href="{{ asset('css/base.css') }}">
+        {% endblock %}
+
     </head>
+
     <body>
-        {% block body %}{% endblock %}
-        {% block javascripts %}{% endblock %}
+
+        {% block main %}{% endblock main %}
+
+    <body>
+
+        {% block javascripts %}
+
+        <!-- JAVASCRIPT -->
+        <script src="{{ asset('js/base.js') }}"></script>
+            
+        <!-- JQUERY CDN - SLIM VERSION (WITHOUT AJAX) -->
+        <script src="https://code.jquery.com/jquery-3.3.1.slim.min.js" integrity="sha384-q8i/X+965DzO0rT7abK41JStQIAqVgRVzpbzo5smXKp4YfRvH+8abtTE1Pi6jizo" crossorigin="anonymous"></script>
+
+        <!-- POPPER.JS -->
+        <script src="https://cdnjs.cloudflare.com/ajax/libs/popper.js/1.14.0/umd/popper.min.js" integrity="sha384-cs/chFZiN24E4KMATLdqdvsezGxaGsi4hLGOzlXwp5UZB1LY//20VyM2taTB4QvJ" crossorigin="anonymous"></script>
+
+        <!-- BOOTSTRAP JS -->
+        <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.0/js/bootstrap.min.js" integrity="sha384-uefMccjFJAIv6A+rW+L4AHf99KvxDjWSu1z9VI8SKNVmz4sk7buKt/6v9KI65qnm" crossorigin="anonymous"></script>
+
+        <!-- PAPER KIT 2.0 -->
+        <script src="{{ asset('js/paper-kit.min.js') }}" type="text/javascript"></script>
+
+        {% endblock javascripts %}
+
     </body>
 </html>
